Added proj files
Added main project files.

Please note:
This was in the beginning, a modified version of the asteroids game. The images and stuff used in the One-Page/Sell Sheet were only for proof of concept and to show what the game could be. So I am not rlly sure how this would turn out lol.